requirements.txt
cupy
faiss_cpu
matplotlib
nltk
numpy
torch
transformers
Whoosh
